upgrade docker-debify to use ruby 2.6
Recent changes to one of the gems bundler depends on mean that it no
longer installs cleanly on a ruby older than 2.3.

Just upgrading to 2.3 was an option, I suppose, but it's scheduled for
EOL at the end of March 2019....

So, instead, I upgraded to 2.6. Old versions of the Conjur API and CLI
don't work with 2.6, so I needed to upgrade them, too. This required
changes to the way we retrieve the Conjur variables that hold the
credentials used to access artifactory.