Upgrade Dockerfile OS from Ubuntu 14.04 --> 18.04
Ubuntu 14.04 has entered  the "Extended Secuirty Maintenance" phase
which means it's repos are behind a paywall. Access can be gained via
an Ubuntu Advantage Subscription from Canonical. This commit takes
the alternate route and upgrades the base iamge to Ubuntu 18.04.

The Dockerfile installs Ruby from `ppa:brightbox/ruby-ng` however that
PPA doesn't currently supply ruby2.2 for Ubuntu
18.04[1]. The documentation[2] suggests this combination is available,
so it may be a temporary problem. To work around the problem, ruby is
bumped from 2.2 to 2.3 as 2.3 is the oldest version available for
Ubuntu 18.04.

Related: conjurinc/appliance#1067

[1] https://launchpad.net/~brightbox/+archive/ubuntu/ruby-ng?field.series_filter=bionic
[2] https://www.brightbox.com/docs/ruby/ubuntu/