Change VERSION read for image tags
The version of bash on the Jenkins executors causes the previous
route of reading the version to terminate early on master branch
builds because of reaching EOF on the VERSION file during the read
command.  This fixes the issue and allows images to be pushed.

It is unknown how long this was broken for.