Supply debian distribution to fpm
In jordansissel/fpm@37f56269c331c47758d0010e335f9d12d3094574 the default
distribution was changed from whatever to null.

When debian packages are promoted (from release to customer release)
the package is expanded, then the version is changed and its repacked.
This failed for conjur-ui because the included changelog was deemed
invalid by the re-package process. The reason for is the lack of
distribution field.

In older packages the distribution field was hardcoded to 'whatever'
by fpm. This commit brings back 'whatever' as the default.

In future we can set this to something more useful if it becomes
necessary.

Related: ONYX-29717